feat(datasets) Extract Storage classes from Dataset for reuse (#829)
This should be a no brainer. Now that we have a class that represents the Storage in a dataset we can reuse them instead of having weird cross references between datasets.
This moves the Storage objects out of the dataset objects so they can be reused.
